pyproject.toml
[project]
name = "master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "pydantic-ai>=0.1.11",
    "sentence-transformers>=4.1.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "ipython>=8.36.0",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "yapf>=0.43.0",
]
